+++ Documentation/CHANGELOG.md
@@ -165,6 +165,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added `ReplayResult.IsExactMatch` convenience property for checking exact replay matches
   - Improved `DebugPackage` integrity hash computation to order records consistently for deterministic hashing
 
+#### Changed
+- Added session ID to audit log.
+
 ## [0.3.0-rc.2] 2026-01-07
 
 ### Unity Runtime